periodic delete + bunch of other things

+++ db.go
@@ -2,11 +2,14 @@ package main
 
 import (
 	"database/sql"
+	"errors"
 	"time"
+
+	_ "modernc.org/sqlite"
 )
 
 var initStatement = `
-CREATE TABLE instances(id INTEGER PRIMARY KEY, challenge TEXT, expiry INTEGER);
+CREATE TABLE IF NOT EXISTS instances(id INTEGER PRIMARY KEY, challenge TEXT, expiry INTEGER);
 DELETE FROM instances;
 `
 
@@ -17,7 +20,7 @@ type InstanceRecord struct {
 }
 
 func (in *Instancer) InitDB(file string) error {
-	db, err := sql.Open("sqlite3", file)
+	db, err := sql.Open("sqlite", file)
 	if err != nil {
 		return err
 	}
@@ -30,6 +33,9 @@ func (in *Instancer) InitDB(file string) error {
 }
 
 func (in *Instancer) InsertInstanceRecord(ttl time.Duration, challenge string) (time.Time, error) {
+	if in.db == nil {
+		return time.Time{}, errors.New("db not initialized")
+	}
 	expiry := time.Now().Add(ttl)
 
 	stmt, err := in.db.Prepare("INSERT INTO instances(challenge, expiry) values(?, ?)")
@@ -38,7 +44,7 @@ func (in *Instancer) InsertInstanceRecord(ttl time.Duration, challenge string) (
 	}
 	defer stmt.Close()
 
-	_, err = stmt.Exec(challenge, expiry)
+	_, err = stmt.Exec(challenge, expiry.Unix())
 	if err != nil {
 		return time.Now(), err
 	}
@@ -47,6 +53,9 @@ func (in *Instancer) InsertInstanceRecord(ttl time.Duration, challenge string) (
 }
 
 func (in *Instancer) DeleteInstanceRecord(id int) error {
+	if in.db == nil {
+		return errors.New("db not initialized")
+	}
 	stmt, err := in.db.Prepare("DELETE FROM instances WHERE id = ?")
 	if err != nil {
 		return err
@@ -62,18 +71,24 @@ func (in *Instancer) DeleteInstanceRecord(id int) error {
 }
 
 func (in *Instancer) ReadInstanceRecords() ([]InstanceRecord, error) {
+	if in.db == nil {
+		return nil, errors.New("db not initialized")
+	}
 	rows, err := in.db.Query("SELECT id, challenge, expiry FROM instances")
 	if err != nil {
 		return nil, err
 	}
 	defer rows.Close()
-	records := make([]InstanceRecord, 15)
+	records := make([]InstanceRecord, 0)
 	for rows.Next() {
 		record := InstanceRecord{}
-		err = rows.Scan(&record.id, &record.challenge, &record.expiry)
+		var t int64
+		err = rows.Scan(&record.id, &record.challenge, &t)
 		if err != nil {
 			return records, err
 		}
+		record.expiry = time.Unix(t, 0)
+		records = append(records, record)
 	}
 	err = rows.Err()
 	return records, err
